refactor : modularised global css into component-wise css
