+++ src/components/EmptyDashboard/EmptyDashboard.js
@@ -1,5 +1,5 @@
 import React from "react";
-import empty from "../assets/illustrations/empty.svg";
+import empty from "../../assets/illustrations/empty.svg";
 import { Link } from "react-router-dom";
 
 const EmptyDashboard = () => {
