+++ src/components/Footer/footer.css
@@ -1,0 +1,31 @@
+.footer {
+  background: var(--clr-primary-2);
+  color: var(--clr-white);
+}
+.code {
+  color: var(--clr-red-dark);
+}
+
+.footer-center {
+  padding: 0.75rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.footer p {
+  color: var(--clr-white);
+  margin-bottom: 0.5rem;
+  font-weight: 400;
+}
+
+.footer-icon {
+  color: var(--clr-white);
+  font-size: 1.25rem;
+  margin-right: 2rem;
+}
+
+.footer-icon:hover {
+  color: hsl(42, 100%, 64%);
+}
