+++ src/components/Loading/Loading.js
@@ -1,4 +1,5 @@
 import React from "react";
+import "./loading.css";
 
 const Loading = () => {
   return (
