+++ src/components/Navbar/Navbar.js
@@ -9,6 +9,7 @@ import {
 import { GiHamburgerMenu } from "react-icons/gi";
 import { Link } from "react-router-dom";
 import { useAuth0 } from "@auth0/auth0-react";
+import "./navbar.css";
 
 function Navbar() {
   const [showNavLinks, setShowNavLinks] = useState(false);
